Chnage historicaIOBCOB to Object based SQL actions, other fixes

+++ app/src/main/java/com/hypodiabetic/happ/Profile.java
@@ -15,17 +15,17 @@ public class Profile {
     public static final Integer max_iob = 3;                //maximum amount of non-bolus IOB OpenAPS will ever deliver
     public static final Double dia = 1.5;                   //Duration of Insulin Action (hours)
     public static final String type = "current";            //? live info from pump?
-    public static final Integer current_basal = 0;          //Your current background basal at this moment of time from the pump
-    public static final Integer max_daily_basal = 0;        //?
-    public static final Integer max_basal = 0;              //?
+    public static final Double current_basal = 0D;          //Your current background basal at this moment of time from the pump
+    public static final Double max_daily_basal = 0D;        //?
+    public static final Double max_basal = 0D;              //?
     public static final Integer min_bg = 90;                //low end of BG Target range
     public static final Integer max_bg = 144;               //high end of BG Target range
     public static final Integer carbratio = 15;             //? live info on current carb to insulin ratio?
-    public static final Integer sens = 54;                   //? see getSensitivity below
+    public static final Double sens = 54D;                   //? see getSensitivity below
 
     //carb profile items
     public static final Double getSensitivity = 25D;         //Insulin sensitivity at this time? In US format
-    public static final Double getCarbRatio = 8D;          //Grams of Carbs per 1 Unit of Insulin
+    public static final Double getCarbRatio = 8D;           //Grams of Carbs per 1 Unit of Insulin
     public static final Double getCarbAbsorptionRate = 30D; //Carbs Disgested per hour http://diyps.org/2014/05/29/determining-your-carbohydrate-absorption-rate-diyps-lessons-learned/
